REFACTOR: Moved files for filtering skills into filter folder
