+++ actions/skills/filter/filterNonEmptySkillCategories.ts
