+++ app/blogs/[slug]/page.tsx
@@ -1,6 +1,6 @@
 import getMarkdownFromFileSystem from "@/actions/file-system/getMarkdownFromFileSystem";
-import filterAndGroupSkills from "@/actions/skills/filterAndGroupSkills";
-import filterSkillsByType from "@/actions/skills/filterSkillsByType";
+import filterAndGroupSkills from "@/actions/skills/filter/filterAndGroupSkills";
+import filterSkillsByType from "@/actions/skills/filter/filterSkillsByType";
 import Reader from "@/components/Reader/Reader";
 import SkillTableSection from "@/components/Skills/SkillTableSection";
 import HeadingTwo from "@/components/Text/HeadingTwo";
